requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
cloudinary==1.37.0
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.25.0
